# api/openapi.yaml
openapi: 3.1.0
info:
  title: SMemories API
  version: 0.1.0
  description: >-
    Hand-written contract for the SMemories API. Infrastructure routes live at the
    root; business routes live under /v1 (the web app reaches them as /api/v1/*).
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /v1/auth/register:
    post:
      operationId: register
      security: []
      tags: [auth]
      summary: Create an account and sign in
      description: >-
        Email is trimmed, lower-cased and normalised to NFC; a duplicate (case-insensitive, and
        NFC/NFD spellings are the same address) is `409 email_taken`. Password: 10-128 characters
        and not equal to the email. Display name: 1-100 characters after trimming, no control or
        format characters (see the field). Rate limit: 5 per hour per IP (counted once the
        input is valid). A cookie-less request needs no Origin; one that carries a session cookie
        must satisfy the CSRF origin check.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, display_name]
              properties:
                email:
                  type: string
                  maxLength: 254
                  description: >-
                    Trimmed, lower-cased and normalised to Unicode NFC, so NFC and NFD spellings of one
                    address are the same account (also at login). The local part (before the last `@`)
                    is at most 64 bytes, and the address contains no control (Cc) or format (Cf)
                    characters; otherwise `invalid_email`.
                password:
                  type: string
                  format: password
                  minLength: 10
                  maxLength: 128
                  description: >-
                    Normalised to Unicode NFKC before the length rules (counted in characters after
                    normalisation) and before hashing, so NFC and NFD forms of the same password are equal.
                display_name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  description: >-
                    Normalised to Unicode NFC, then trimmed; stored and returned in that form. Control
                    characters (Cc), format characters (Cf: zero-width space, bidi overrides and isolates,
                    LRM/RLM, BOM) except U+200D (zero-width joiner, used in emoji sequences), and line or
                    paragraph separators (Zl, Zp) are rejected with `invalid_display_name`.
      responses:
        "201":
          description: Account created; the session cookie is set.
          headers:
            Set-Cookie: {$ref: "#/components/headers/SessionCookie"}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/UserEnvelope"}
        "400":
          description: >-
            `invalid_email`, `weak_password`, `invalid_display_name` or `invalid_body`.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "403": {$ref: "#/components/responses/CsrfOriginMismatch"}
        "409":
          description: Email already registered (`email_taken`).
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "413": {$ref: "#/components/responses/PayloadTooLarge"}
        "415": {$ref: "#/components/responses/UnsupportedMediaType"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "503": {$ref: "#/components/responses/Busy"}
  /v1/auth/login:
    post:
      operationId: login
      security: []
      tags: [auth]
      summary: Sign in with email and password
      description: >-
        A wrong password and an unknown email give the identical `401 invalid_credentials`
        (same body, same work). Every login issues a new session token; a session cookie sent
        with the request is revoked. Failures are limited to 10 per 15 minutes per IP+email and
        100 per 15 minutes per IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: {type: string, description: "Trimmed, lower-cased and NFC-normalised as at registration, so either Unicode spelling signs in."}
                password:
                  type: string
                  format: password
                  description: Normalised to Unicode NFKC before verification, exactly as at registration.
      responses:
        "200":
          description: Signed in; the session cookie is set.
          headers:
            Set-Cookie: {$ref: "#/components/headers/SessionCookie"}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/UserEnvelope"}
        "400":
          description: Malformed body (`invalid_body`).
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "401":
          description: Wrong email or password (`invalid_credentials`).
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "403": {$ref: "#/components/responses/CsrfOriginMismatch"}
        "413": {$ref: "#/components/responses/PayloadTooLarge"}
        "415": {$ref: "#/components/responses/UnsupportedMediaType"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "503": {$ref: "#/components/responses/Busy"}
  /v1/auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Sign out (idempotent)
      description: Deletes the session and clears the cookie. Always 204, also when nobody is signed in.
      security:
        - sessionCookie: []
        - {}
      responses:
        "204":
          description: Signed out; the cookie is cleared.
          headers:
            Set-Cookie:
              description: smem_session with an empty value and a negative Max-Age.
              schema: {type: string}
        "403": {$ref: "#/components/responses/CsrfOriginMismatch"}
        "415": {$ref: "#/components/responses/UnsupportedMediaType"}
  /v1/auth/verify-email:
    post:
      operationId: verifyEmail
      security: []
      tags: [auth]
      summary: Confirm an email address with the token from the verification email
      description: >-
        Tokens are 32 random bytes (base64url), stored only as SHA-256, valid 24 hours and single use.
        An unknown, expired, used or wrong-purpose token (a reset token, for example) is `400 invalid_token`.
        Registration and `verify-email/resend` send the email; the link is
        `SMEM_PUBLIC_BASE_URL/verify-email?token=...`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: {type: string}
      responses:
        "204":
          description: The address is verified (`email_verified` is now true).
        "400":
          description: "`invalid_token` or `invalid_body`."
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "403": {$ref: "#/components/responses/CsrfOriginMismatch"}
        "413": {$ref: "#/components/responses/PayloadTooLarge"}
        "415": {$ref: "#/components/responses/UnsupportedMediaType"}
  /v1/auth/verify-email/resend:
    post:
      operationId: resendVerificationEmail
      tags: [auth]
      summary: Send a new verification email to the signed-in user
      description: >-
        At most 3 per hour per user (`429 rate_limited` with `Retry-After`). Earlier links stay valid
        until they expire. Takes no body.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: The address was already verified; nothing was sent.
          content:
            application/json:
              schema:
                type: object
                required: [already_verified]
                properties:
                  already_verified: {type: boolean, const: true}
        "202":
          description: The email was sent. A mailer failure is `500 internal_error` (and is logged without the token).
          content:
            application/json:
              schema: {type: object}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/CsrfOriginMismatch"}
        "415": {$ref: "#/components/responses/UnsupportedMediaType"}
        "429": {$ref: "#/components/responses/RateLimited"}
  /v1/auth/forgot-password:
    post:
      operationId: forgotPassword
      security: []
      tags: [auth]
      summary: Ask for a password reset email
      description: >-
        Always `202 {}` for a well-formed address, whether or not an account exists; the email (valid
        link for 1 hour, in the user's `locale`) goes out in the background and only when the account
        exists, so neither the body nor the timing tells them apart. Limits: 5 per hour per IP and 3 per
        hour per address (counted for unknown addresses too). The address is trimmed, lower-cased and
        NFC-normalised like at login; one that is not an email at all is `400 invalid_email`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: {type: string, maxLength: 254}
      responses:
        "202":
          description: Accepted.
          content:
            application/json:
              schema: {type: object}
        "400":
          description: "`invalid_email` or `invalid_body`."
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "403": {$ref: "#/components/responses/CsrfOriginMismatch"}
        "413": {$ref: "#/components/responses/PayloadTooLarge"}
        "415": {$ref: "#/components/responses/UnsupportedMediaType"}
        "429": {$ref: "#/components/responses/RateLimited"}
  /v1/auth/reset-password:
    post:
      operationId: resetPassword
      security: []
      tags: [auth]
      summary: Set a new password with the token from the reset email
      description: >-
        The token is valid 1 hour and single use; an unknown, expired, used or wrong-purpose token is
        `400 invalid_token`. The password follows the registration rules (NFKC, 10-128 characters, not the
        email); a rejected password is `400 weak_password` and does not use up the token. On success every
        session of the user is deleted (sign in again), and the user's other open reset links stop working.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: {type: string}
                password: {type: string, format: password, minLength: 10, maxLength: 128}
      responses:
        "204":
          description: Password changed; all sessions are gone.
        "400":
          description: "`invalid_token`, `weak_password` or `invalid_body`."
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "403": {$ref: "#/components/responses/CsrfOriginMismatch"}
        "413": {$ref: "#/components/responses/PayloadTooLarge"}
        "415": {$ref: "#/components/responses/UnsupportedMediaType"}
        "503": {$ref: "#/components/responses/Busy"}
  /v1/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: The signed-in user
      description: >-
        Sliding expiry: once more than half of the 30-day lifetime has elapsed the expiry is
        extended and the cookie is sent again.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: The current user.
          headers:
            Set-Cookie:
              description: Present only when the session expiry was extended.
              schema: {type: string}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/UserEnvelope"}
        "401":
          description: No, unknown or expired session (`unauthenticated`).
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: smem_session
      description: >-
        HttpOnly, SameSite=Lax, Path=/, Max-Age 30 days, Secure except when SMEM_ENV=dev.
        State-changing requests that carry it must come from an origin in SMEM_ALLOWED_ORIGINS.
  headers:
    SessionCookie:
      description: >-
        `smem_session=<32 random bytes, base64url>; Path=/; Max-Age=2592000; HttpOnly; SameSite=Lax`
        (plus `Secure` unless SMEM_ENV=dev). Only the token's SHA-256 is stored server-side.
      schema: {type: string}
  responses:
    Unauthenticated:
      description: No, unknown or expired session (`unauthenticated`).
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
...
    CsrfOriginMismatch:
      description: >-
        The request carries the session cookie but its Origin (or Referer) is missing or not
        allowed (`csrf_origin_mismatch`).
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    PayloadTooLarge:
      description: Body over the 1 MiB cap (`payload_too_large`).
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    UnsupportedMediaType:
      description: A body was sent with a Content-Type other than application/json (`unsupported_media_type`).
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    RateLimited:
      description: Too many attempts (`rate_limited`).
      headers:
        Retry-After:
          description: Seconds until the oldest counted attempt leaves the window.
          schema: {type: integer, minimum: 1}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
...
    Busy:
      description: >-
        All password-hashing slots stayed taken for 2 seconds (`busy`); retry shortly.
      headers:
        Retry-After:
          schema: {type: integer}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
...
  schemas:
    User:
      type: object
      additionalProperties: false
      required: [id, email, email_verified, display_name, locale, created_at]
      properties:
        id: {type: string, description: Opaque ULID., examples: ["01J9Z3K6V8Q4M7N2P5R8T0W1XY"]}
        email: {type: string}
        email_verified: {type: boolean}
        display_name: {type: string}
        locale: {type: string, enum: [en, vi]}
        created_at: {type: string, format: date-time}
    UserEnvelope:
      type: object
      required: [user]
      properties:
        user: {$ref: "#/components/schemas/User"}
...
    Error:
      type: object
      description: >-
        Shared error envelope. `code` is a stable snake_case contract (clients localise
        it); `message` is English text meant for logs.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              examples: [not_found, method_not_allowed, payload_too_large, internal_error, not_ready, invalid_credentials, unauthenticated, rate_limited, invalid_token, limit_reached, unknown_field, unsupported_media_type, invalid_image, quota_exceeded, storage_error, invalid_media]
            message:
              type: string
            request_id:
              type: string
              description: Same value as the X-Request-Id response header.